Extract TA functionality into own class
