+++ requirements.txt
@@ -25,11 +25,14 @@ packaging==16.8
 ply==3.8
 proto-google-cloud-datastore-v1==0.90.4
 protobuf==3.3.0
+py==1.4.34
 pyasn1==0.2.3
 pyasn1-modules==0.0.8
 pyparsing==2.2.0
+pytest==3.1.2
 requests==2.13.0
 rsa==3.4.2
 six==1.10.0
+uritemplate==3.0.0
 urllib3==1.21.1
 Werkzeug==0.12.1
